requirements.txt
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
lxml==5.2.1
oauthlib==3.2.2
packaging==24.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
